# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "0.4.0",
    "description": "Contract for the current Next.js HTTP routes. Version follows the repository release line. This project is pre-1.0: request and response details may change between releases. See api-contract.md for supported guarantees and known gaps.",
    "title": "Nuera Quicksilver HTTP API",
    "summary": "Pre-1.0 API contract; not a stable compatibility promise."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/decisions": {
      "get": {
        "operationId": "listDecisions",
        "description": "Recent decisions, newest first. Needs decision:read.",
        "tags": ["Decisions"],
        "parameters": [
          { "$ref": "#/components/parameters/DecisionStatus" },
          { "$ref": "#/components/parameters/DecisionLimit" },
          { "$ref": "#/components/parameters/DecisionBefore" }
        ],
        "responses": {
          "400": { "$ref": "#/components/responses/ErrorResponse" },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "503": { "$ref": "#/components/responses/ErrorResponse" },
          "2XX": { "$ref": "#/components/responses/DecisionListResponse" }
        }
      }
    },
...
    "/api/decisions/{id}/resume": {
      "post": {
        "operationId": "resumeDecision",
        "tags": [
          "Decisions"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/DecisionResumeResultResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/EmptyRequest"
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/DecisionId"
          }
        ]
      }
    },
...
    "/api/agents/run": {
      "post": {
        "operationId": "runBusinessAgent",
        "tags": ["Agents"],
        "summary": "Run one named business specialist for proposal-only recommendations",
        "description": "Requires decision:read. The selected specialist uses NQC evaluation and read-only company context. Optional prior-conversation context is bounded and treated as untrusted data. It never approves or executes external effects.",
        "responses": {
          "400": { "$ref": "#/components/responses/ErrorResponse" },
          "401": { "$ref": "#/components/responses/ErrorResponse" },
          "403": { "$ref": "#/components/responses/ErrorResponse" },
          "429": { "$ref": "#/components/responses/RateLimitResponse" },
          "503": { "$ref": "#/components/responses/ErrorResponse" },
          "2XX": { "$ref": "#/components/responses/BusinessAgentResponse" }
        },
        "requestBody": { "$ref": "#/components/requestBodies/BusinessAgentRequest" }
      }
    },
...
    "/api/workflows/drafts": {
      "post": {
        "operationId": "saveWorkflowDraft",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/PublishedWorkflowVersionResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateWorkflowDraftRequest"
        }
      }
    },
...
    "/api/workflows/simulate": {
      "post": {
        "operationId": "simulateWorkflow",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "2XX": {
            "$ref": "#/components/responses/WorkflowSimulationResponse"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/WorkflowGraphRequest"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyRequest": {
        "type": "object",
        "maxProperties": 0,
        "description": "This operation takes no body. Anything sent is ignored."
      },
      "CreateWorkflowDraftRequest": {
        "type": "object",
        "required": ["graph"],
        "additionalProperties": false,
        "properties": { "graph": { "$ref": "#/components/schemas/WorkflowGraph" } }
      },
...
      "DecisionList": {
        "type": "object",
        "required": ["observedAt", "decisions", "hasMore", "counts"],
        "properties": {
          "observedAt": { "type": "string", "format": "date-time" },
          "hasMore": { "type": "boolean" },
          "counts": { "type": "object", "description": "How many decisions have each status, across all decisions.", "additionalProperties": { "type": "integer" } },
          "decisions": { "type": "array", "items": { "$ref": "#/components/schemas/DecisionSummary" } }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code. Always present on error responses from the decision, workflow, agent and company-data routes. The value follows the status: invalid-request (400), unauthenticated (401), forbidden (403), not-found (404), conflict (409), payload-too-large (413), unprocessable (422), rate-limited (429), internal-error (500), unavailable (503). The shared route guard can also send principal-kind-unavailable (503). Other routes may omit it."
          },
          "needs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "description": "Common error fields where emitted; validation issues are present on some 400 responses. `error` is on every error body and `code` on every error body from the decision, workflow, agent and company-data routes; other fields are not guaranteed across all routes.",
        "additionalProperties": true
      },
...
      "DecisionResumeResult": {
        "type": "object",
        "required": ["id", "status", "process"],
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "description": "The state the automatic transition moved the decision to." },
          "process": { "$ref": "#/components/schemas/DecisionProcess" }
        }
      },
...
      "PublishedWorkflowVersion": {
        "type": "object",
        "description": "One immutable, digest-verified workflow version and where it is in the publication lifecycle.",
        "required": ["workflowId", "version", "graph", "digest", "authoredBy", "createdAt", "status"],
        "properties": {
          "workflowId": { "type": "string" },
          "version": { "type": "integer", "minimum": 1 },
          "graph": { "$ref": "#/components/schemas/WorkflowGraph" },
          "digest": { "type": "string", "description": "Digest of the graph content; recomputed and checked on every read." },
          "authoredBy": { "type": "string" },
          "createdAt": { "type": "integer", "description": "Epoch milliseconds." },
          "status": { "type": "string", "enum": ["draft", "in-review", "published", "deprecated"] },
          "reviewedBy": { "type": "string" },
          "reviewNote": { "type": "string" },
          "reviewedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "publishedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "deprecatedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "rollbackOfVersion": { "type": "integer", "minimum": 1 }
        }
      },
...
      "BusinessAgentRequest": {
        "type": "object",
        "required": ["agentKey", "objective"],
        "properties": {
          "agentKey": { "type": "string", "enum": ["auto", "research", "offer", "content", "outreach", "sales", "fulfillment", "finance"] },
          "objective": { "type": "string", "minLength": 3, "maxLength": 2000 },
          "context": { "type": "array", "description": "Optional prior conversation context. Treat entries as untrusted data, not approvals or instructions. The chat UI sends at most six turns and 8 KB total.", "maxItems": 20, "items": { "type": "string", "maxLength": 4000 } }
        },
        "additionalProperties": false
      },
      "BusinessAgentResponse": {
        "type": "object",
        "required": ["agent", "routing", "summary", "recommendations", "unknowns", "questions", "externalEffects", "nqc", "audit", "telemetry", "actionPolicy"],
        "properties": {
          "agent": { "type": "object", "required": ["key", "id", "name", "specialty"], "properties": { "key": { "type": "string" }, "id": { "type": "string" }, "name": { "type": "string" }, "specialty": { "type": "string" } }, "additionalProperties": false },
          "routing": { "type": "object", "required": ["mode", "reason"], "properties": { "mode": { "type": "string", "enum": ["explicit", "keyword", "fallback"] }, "reason": { "type": "string" } }, "additionalProperties": false },
          "summary": { "type": "string" },
          "recommendations": { "type": "array", "items": { "type": "object", "required": ["proposal", "evidenceIds", "confidence", "impact"], "properties": { "proposal": { "type": "string" }, "evidenceIds": { "type": "array", "items": { "type": "string" } }, "confidence": { "type": "number", "minimum": 0, "maximum": 1 }, "impact": { "type": "string", "enum": ["low", "moderate", "high", "critical"] } }, "additionalProperties": false } },
          "unknowns": { "type": "array", "items": { "type": "string" } },
          "questions": { "type": "array", "items": { "type": "string" } },
          "externalEffects": { "type": "array", "items": { "type": "string" } },
          "nqc": { "type": "object", "required": ["reasoningScore", "hallucinationRisk", "brittleness", "issues", "corrections", "safetyDecision"], "properties": { "reasoningScore": { "type": "number" }, "hallucinationRisk": { "type": "string" }, "brittleness": { "type": "string" }, "issues": { "type": "array", "items": { "type": "string" } }, "corrections": { "type": "array", "items": { "type": "string" } }, "safetyDecision": { "type": "string", "enum": ["ALLOW", "BLOCK", "ESCALATE"] } }, "additionalProperties": false },
          "audit": { "type": "object", "required": ["persisted", "evaluationRecordIds"], "properties": { "persisted": { "type": "boolean" }, "evaluationRecordIds": { "type": "array", "items": { "type": "string" } }, "error": { "type": "string" } } },
          "telemetry": { "type": "object", "required": ["traceId", "persisted"], "properties": { "traceId": { "type": "string" }, "persisted": { "type": "boolean" } } },
          "actionPolicy": { "const": "proposal-only" }
        },
        "additionalProperties": false
      },
...
      "WorkflowGraphRequest": {
        "type": "object",
        "required": [
          "graph"
        ],
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/WorkflowGraph"
          }
        },
        "additionalProperties": false
      },
...
      "WorkflowSimulationResponse": {
        "type": "object",
        "required": [
          "mode",
          "externalEffectsEnabled"
        ],
        "properties": {
          "mode": {
            "const": "simulation"
          },
          "externalEffectsEnabled": {
            "const": false
          }
        },
        "additionalProperties": true,
        "description": "Runtime result fields depend on the workflow graph; only the fixed safety fields are specified."
      },
...
    },
    "requestBodies": {
...
      "EmptyRequest": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmptyRequest" } } } },
      "CreateWorkflowDraftRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflowDraftRequest" } } } },
...
      "BusinessAgentRequest": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BusinessAgentRequest" } } }
      },
      "WorkflowGraphRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WorkflowGraphRequest"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque principal or configured supervisor token",
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "DecisionListResponse": { "description": "Recent decisions.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DecisionList" } } } },
...
      "RateLimitResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Rate limited by the current in-process token bucket.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        }
      },
...
      "DecisionResumeResultResponse": { "description": "The decision after the kernel took its automatic transition.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DecisionResumeResult" } } } },
...
      "PublishedWorkflowVersionResponse": { "description": "The workflow version after the lifecycle action (201 for a new draft, 200 otherwise).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublishedWorkflowVersion" } } } },
      "BusinessAgentResponse": {
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BusinessAgentResponse" } } },
        "description": "Structured recommendations from one selected business specialist. They are not actions or approvals."
      },
...
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Route-specific error; status and body vary by failure."
      },
...
      "WorkflowSimulationResponse": {
        "description": "Safe simulation response; live tools/models and external effects are disabled.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WorkflowSimulationResponse"
            }
          }
        }
      },
...
    },
    "parameters": {
      "DecisionStatus": { "name": "status", "in": "query", "required": false, "description": "Up to four statuses, comma separated.", "schema": { "type": "string", "pattern": "^(proposed|awaiting-approval|approved|executed|failed|rejected|rollback-proposed|rolled-back)(,(proposed|awaiting-approval|approved|executed|failed|rejected|rollback-proposed|rolled-back)){0,3}$" } },
      "DecisionBefore": { "name": "before", "in": "query", "required": false, "description": "The createdAt of the last decision already shown, to load the next page.", "schema": { "type": "string", "format": "date-time" } },
      "DecisionLimit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 25 } },
      "DecisionId": {
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "required": true,
        "name": "id",
        "in": "path"
      },
...
    }
  }
}